backend/requirements.txt
pandas
fastapi[all]
uvicorn
db-sqlite3
pygithub
uvicorn
sqlalchemy
fire
